+++ frontend/src/lib/components/universe/SelectProjectDropdown.svelte
